refactor(bot): moved media classes to tg
