+++ src/kamihi/bot/action.py
@@ -18,8 +18,8 @@ from telegram import Update
 from telegram.constants import BotCommandLimit
 from telegram.ext import ApplicationHandlerStop, CallbackContext, CommandHandler
 
+from kamihi.tg import send
 from kamihi.tg.handlers import AuthHandler
-from kamihi.tg.send import send
 from kamihi.users import get_user_from_telegram_id
 
 from .models import RegisteredAction
