+++ src/kamihi/bot/bot.py
@@ -29,10 +29,10 @@ from telegram import BotCommand
 from telegram.ext import CommandHandler
 
 from kamihi.base.config import KamihiSettings
-from kamihi.bot.media import Audio, Document, Location, Photo, Video
 from kamihi.db.mongo import connect, disconnect
 from kamihi.templates import Templates
 from kamihi.tg import TelegramClient
+from kamihi.tg.media import *
 from kamihi.users import get_users, is_user_authorized
 from kamihi.users.models import User
 from kamihi.web import KamihiWeb
